requirements.txt
flask>=3.0
flask-session>=0.8
exifread>=3.0
Pillow>=10.0
rawpy>=0.18
numpy>=1.24
anthropic>=0.25
python-dotenv>=1.0
